# docs/swagger.yaml
basePath: /
definitions:
  handler.CreateTopicRequest:
    properties:
      numPartitions:
        example: 3
        minimum: 1
        type: integer
      replicationFactor:
        example: 1
        minimum: 1
        type: integer
    required:
    - numPartitions
    - replicationFactor
    type: object
  handler.MessageRequest:
    properties:
      key:
        example: user-123
        type: string
      value:
        example: Hello, Kafka!
        type: string
    required:
    - value
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A REST API gateway for Apache Kafka operations
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Kafka Gateway API
  version: "1.0"
paths:
  /api/v1/publish/{topic}:
    post:
      consumes:
      - application/json
      description: Publish a message to a specified Kafka topic
      parameters:
      - description: Topic name
        in: path
        name: topic
        required: true
        type: string
      - description: Message to publish
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/handler.MessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Publish message to Kafka topic
      tags:
      - kafka
  /api/v1/topics:
    get:
      description: Get a list of all available Kafka topics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all Kafka topics
      tags:
      - kafka
  /api/v1/topics/{topic}:
    post:
      consumes:
      - application/json
      description: Create a new topic with specified partitions and replication factor
      parameters:
      - description: Topic name
        in: path
        name: topic
        required: true
        type: string
      - description: Topic configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateTopicRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new Kafka topic
      tags:
      - kafka
  /api/v1/topics/{topic}/partitions:
    get:
      description: Get partition information for a specific Kafka topic
      parameters:
      - description: Topic name
        in: path
        name: topic
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get topic partitions
      tags:
      - kafka
  /health:
    get:
      description: Get the health status of the service
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check endpoint
      tags:
      - health
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"